Create the base class for define the interface for check should able to modify the interface or not.
